Work in progress on making reader part of the stream instead of the receiver
The receiver class is generalised to spead::thread_pool.

There are still some fundamental design issues around stopping: at
present it could deadlock, because a reader calls stop, which dispatches
another callback and then waits for it. With only one thread in the
thread pool, the callback can never run. This probably needs stream or
stream_base to have a separate stop_async function for use by the
reader.