Fix up the Python bindings for API changes
There are still some fairly serious thread safety issues: the GIL is
dropped in a number of places because not doing so leads to deadlocks
when strand-holders try to acquire it (for logging) while the main
thread is blocking. But this makes the code potentially racy if two
Python threads try to access the same stream. There probably needs to be
a regular mutex to protect the list of readers (unless it can be
entirely protected by the strand?)

buffer_reader will also not work, as it is passing a py::object between
threads, which will no doubt cause all kinds of refcounting interaction
with the GIL.