Rewrite the descriptor encoding to operate directly.
This gives us direct control over the addressing mode, and reduces the
number of memory allocations and copies running around. The logic for
immediates is changed to allow padding, since otherwise PySPEAD can't
receive small items. Transmission is still not fully working, because
PySPEAD assumes that immediates have the padding in front.