Fix a race condition
If the destructor for a send stream was called after send_next_packet
popped the last heap and immediately after it dropped the lock (but
before signalling the condition variable) the stream would be destroyed
while it was still being touched.

All changes to the object, as well as the signalling of the condition
variable, are now done with the lock held.