Don't adjust send_time when we've fallen behind
Falling behind can easily happen if we oversleep. This allows us to
catch up from an oversleep. Since send_time is reset every time a heap
arrives when we have an empty queue, it won't allow for huge bursts if
there is a gap in production.

This brings the actual rate much closer to the requested rate in
bandwidth tests, making it possible to hit higher transmission speeds in
transmission.