Use blocking push for ready_heaps
This prevents heaps being dropped merely because the consumer isn't
keeping up. Instead, this will block the asio threads that are reading
from the network, eventually causing packets to be dropped.

The ringbuffer size is now also fixed at 2, rather than depending on
max_heaps. This should reduce memory usage.

Documentation still needs to be updated for this change.