Revert "Use blocking push for ready_heaps"
The design was flawed as regards stop(). If the consumer doesn't drain
the ringbuffer, it will fill up, which blocks heap_ready, which is
holding the strand, which prevents stop_impl from getting it. So a
consumer could easily deadlock in its call to stop().

It also broke Python, because the space_sem was semaphore_gil instead of
just semaphore.

The only solution I can think of is to make space_sem a semaphore_fd,
and then do an asynchronous wait (using asio) for space to become
available. That will require an interface change to the readers (for the
stream to pause the readers when blocked and restart them when
unblocked). It may also require ringbuffer (and hence semaphore) to have
more stringent thread safety so that the asynchronous push can happen
concurrently with stop().

This reverts commit 035d0fd40110801b063f595a2cfe0460c07951de.