Simplifications for memory pool pointers
Instead of the destructor being a std::function, it is now directly an
object (avoids some malloc calls). It also now holds a strong instead of a
weak pointer. Originally I used weak pointers because I was worried about
reference cycles, but only allocated buffers hold the reference and so the
pool is correctly destroyed once all its pointers are allocated.

This makes a small improvement to the small-heap case.