Rework the way duplicate packets are detected
Previously the heap offsets were kept in a hash table. That performed
poorly with small packets, as it required memory allocation per packet,
and hash table operations aren't very fast. The new approach uses a
std::map to store contiguous ranges. In the worst case (random packet
ordering) it will be much slower, but for a small number of in-order
streams it should be much faster.

It still needs some testing with out-of-order packets.