Fix Python refcounting issue with memory pool containing thread pool
The memory pool for a stream was only being refcounted at the C++ level.
The Python object could thus disappear if the user didn't keep an
explicit reference, causing the custodian_and_ward relationship to be
lost.

This change makes the Python MemoryPool object hang around until the
stream dies.