Rewrite memory allocation to put metadata in the pointer type
For some reason the approach of putting data in the deleter was causing
failures in GCC release builds (but not clang or GCC debug builds). It
seems like the deleter and the pointer aren't necessarily glued together
- particularly in release() - although the exact cause of the bug was
not found.

The new approach overrides the pointer type to memory_allocation, which
bundles a regular pointer, a shared_ptr to the allocator, and the user
data. This makes the pointer a lot more heavy-weight, so performance
considerations should be checked.