Batch writes in 8MB chunks.
This will make it easier to support O_DIRECT in the near future. It is
still unknown whether it is best to do the memory copies in the network
thread, the disk thread, or even to use a thread-pooling approaches to
allow for more threads to cooperate on the job.