Optimise Python bindings with SPEAD2_PTMF
Instead of passing pointers to member functions, which requires
cpp_function to create stateful lambdas and do run-time lookup of the
function pointers, a SPEAD2_PTMF macro generates a function object which
bakes in the choice of member function at compile time. Code size
increases as a result, but overheads should be fractionally reduced.