Add incomplete_heap class
This makes it possible to extract some information from incomplete
live_heaps. In particular, it gives access to the ranges of payload that
were received, which is useful for applications which use a custom
memory allocator to direct the payload into custom storage, and need to
know which parts of that storage now contain valid data.