Redesign logging thread to not use gil_scoped_release
This fixes the crash on shutdown in Python 3. Instead of starting by
acquiring the GIL, then releasing it when blocking, the thread now only
acquires the GIL when it needs it. To make this work, ownership of
Python objects now rests with the class itself, and they're cleaned up
by the main thread.