Initial TCP stream sender implementation
This follows very closely the UDP sender implementation, except that the
ttl-based and multicast-capable constructors are gone. Also, because the
UDP transport maps SPEAD packets one-to-one to UDP datagrams, we need to
send the SPEAD packet size via TCP before sending the data itself
(otherwise there is no way to find the end of the stream).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>