Reimplementing TCP receiver with single buffer
This follows the suggestion from Bruce where instead of a
double-buffered approach we use a single buffer with a head and a tail
to receive data and interpret it. The code is much easier to read and
follow, and the state machine of the class much simpler, since it
contains less properties that need to be tracked.

Early experiments seemed to indicate that this implementation was
running slower than the double-buffered alternative (which seemed
counter-intuitive), but further testing revealed that no noticeable
performance loss can really be observed in tightly controlled
environments. For instance, I was getting ~37 [Gb/s] in one of our MRO
machines, while Bruce is getting speeds of ~46 [Gb/s] in his testing
system.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>