New get_packet_size function
This new function takes a number of bytes, tries to interpret it as
SPEAD packet content, and returns the total size of that packet,
including its payload.

The amount of checks and data decoding done by get_packet_size is
similar, simpler and smaller than decode_packet, but not quite the same;
therefore it needs its own bit of logic. The only very common bit of
functionality seems to be the decoding of the SPEAD header (first 8
bytes of data), so a new decode_spead_header function has been extracted
to do exactly this.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>