recv buffer size needs to be set on acceptor
The very first version of this code did set the receiver buffer size on
the acceptor. I was pointed out, however, that this might not be
reliable, since not all settings on the acceptor could be inherited by
the data socket. There were some pointers suggesting this [1], but
empirical evidence in our 100 [Gb/s] link showed that the acceptor
indeed needs to be the one having its receiver buffer size set for the
buffer sizes to be correctly set, and connectivity to work at the
desired speeds. Reading more closely, the inheritance talked about in
[1] does not refer to socket options but to file flags, which is more
clearly pointed out in [2], and the tcp manpage clearly states that the
option has to be set before listen.

[1] https://stackoverflow.com/questions/5968132/are-socket-options-inherited-across-accept-from-the-listening-socket/12864681#12864681
[2] https://stackoverflow.com/questions/39771804/setting-buffer-size-of-accepted-sockets

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>