No need for using a shared_ptr
I based this code on that from the async_send_heap function on the same
module, which indeed *does* require a shared pointer to be used because
the promise quickly goes out of scope. In my case this isn't the case,
so it is safe to use a reference instead.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>