Tracking if the TCP socket is connected
This allows us to return a specific error in case users try to push data
through this tcp_stream but the connection hasn't been established yet.
Without this specific handling the stream usually fails with a EPIPE
error, which would be indistinguishable, for instance, from situations
when the remote end closes the connection unexpectedly.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>