Bringing back io_service_ref + socket constructor
This is to cater for the use case when a specific thread pool needs to
be used with the stream, which will be wrapped inside the
io_service_ref object. The io_service itself still needs to be the same
as the socket's, so we are adding back that check (and including the
stdexcept header only in the .cpp file).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>