New TCP transport defaults
The default packet size on the receiver side has gone up now to 65536
(i.e., the MTU on a loopback interface). This is because TCP already
fully packs user data into frames, and therefore we only loose when
using small packet sizes. On the sender side the default maximum packet
size is not a member of the tcp/udp_stream class, but of the
stream_config class. This is exposed all the way up to python, so I am a
bit hesitant to change that at the moment.

Additionally, the buffer sizes do not need to be so large as in UDP's
case given TCP's nature. Their new values are based on the default
maximum buffer sizes allowed in Linux (208 Kb) both for sending and
receiving.

Finally, the pkts_per_buffer was always meant to be a private member,
which is now.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>