Handle "too many items in packet" case gracefully
If too many items are contained within an incoming packet, the routine
that parses the item size will keep saying that more data is needed. In
time the TCP reader buffer will fill up, and since the head pointer
would not move at all the memmove routine would not be called, and an
async read pointing to memory after the reader buffer's end would be
issued.

This change deals with a way to give up when more data is needed by the
packet size parser, but no more data can be fed into it. In that case we
discard the whole reader buffer and start anew, hoping that at some
point a new packet with a supported size will be seen in the stream.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>