Make constructors with custom socket take pre-configured socket
For UDP these add new constructors, and the existing constructors are
deprecated. For TCP the send constructor is replaced, and the receive
constructor was already like this.

The documentation still needs to be updated to match, and the TCP Python
bindings still need to be completed.

There is a lot of internal change so that sockets now use the pybind11
custom converter machinery, so that overload resolution can tell whether
something looks like a socket (which will make overload resolution less
ambiguous).