Change udp_stream back to taking an unconnected socket
Taking a connected socket but still using sendto causes errors on OSX.
The "new" socket constructor now takes an endpoint. This unfortunately
makes it ambiguous with the old constructor when no buffer_size is
given. The new constructor is used in that case (or equivalently,
buffer_size is 0 rather than the default).

In Python, the constructors are adjusted so that the new constructor
takes the socket parameter earlier (more in line with the C++
constructor). There is a slight difference in that socket=None will no
longer work.