Fix race condition in initialisation tcp_stream
The connect call was being made early in the constructor, and in some
cases would complete and fire the handler before the constructor had
finished initialising the class. That in turn would cause the connected
flag to be set back to zero by the constructor.

The connect call is now only made at the end of the constructor body,
and the constructors are someone reworked (in particular, the
constructor that doesn't take an existing socket no longer chains to the
one that does).