Store pointers in live_heap inline
Up to 8 pointers are stored in the heap object itself. The main
advantage is actually eliminating the use of seen_pointers (a set) when
in the few-pointers regime.

This still needs some targeted tests to ensure that the switch between
regimes is handled correctly, and it may be beneficial to force the
switch earlier if the current packet has too many items for the inline
regime.