Batching for send::udp_ibv_stream
It now sucks in up to n_slots packets before sending anything. This
probably needs to be refined to avoid being too lumpy.

Also fixes a race condition: if we had insufficient slots available and
then requested a completion notification, it's possible that all pending
work completed before the request, in which case it would never arrive
and we would deadlock.