Optimise spead2_send.cpp
The promise/future-based synchronisation is replaced by doing the work
in the asio callbacks directly. This has much lower overhead.
Additionally, the heaps are all pre-constructed. The run function is
turned into a class to facilitate callbacks that don't need to bind a
lot of state.

The --threads option is removed because multiple threads are never
actually useful, and with the current round-robin design can lead to
heaps getting arbitrarily out-of-order as the threads progress at
different rates.