spead2_send/spead2_bench: ensure in-order delivery
There was a theoretical possibility that a receive callback could run
before the initial set of heaps was passed to async_send_heap. That
would have caused subsequent heaps to have been sent out of order. That
in turn could have led to the end-of-stream heap being sent too early.