Support dynamic address in asyncio UdpStream
I did some major refactoring of the asyncio module in the process. Most
of the code is now in a helper class (`_AsyncHelper`) rather than put
into dynamically-generated classes.

async_send_heap is now a coroutine rather than a function returning a
future - it originally returned a future because Trollius didn't have
the same semantics as native Python async/await (in particular,
`await foo()` will immediate run code in `foo` until it blocks).

This means that the stream classes are now plain vanilla classes that
should be more amenable to static analysis. Only the docstrings are
manipulated by "magic".