Tweaks to parsing of --bind for recv
Now that --ibv can handle unicast, --bind is not just for multicast. If
--bind is given and a source is given without a host (just a port
number), the bind address is used as the IP address. It is still
possible to override the IP address: in this case the --bind option
determines which physical interface is used, while the IP address is
used to match the packets. This may have utility as an interface is
permitted to have multiple IP addresses.

Note that this also affects non-ibverbs paths.