Wire up send substreams to Python
There are no unit tests yet.

Also fix some undefined behaviour due to over-aggressive use of
std::move in forwarding constructors, which could potentially destroy an
object before it gets used. It's working up to now, presumably because
of the order in which GCC is constructing the arguments to functions.