Improve type inference in send stream constructors
With constructors taking a vector of endpoints or a (deprecated) single
endpoint, passing `{endpoint}` would prefer the single constructor,
treating the braces as constructor arguments rather than an initializer
list. Added explicit initializer_list constructors forwarding to the
vector versions so that this syntax will not lead to deprecation
warnings.