Port TCP send to send::writer framework
Modify the framework to abstract the startup process. By default it
calls request_wakeup, but for TCP we want to connect before we even try
to get packets.

This is an improvement over the previous design, which relied on an
atomic (that I was never totally happy with). It's now guaranteed that
tcp_writer will only have one callback pending at any time. It also
makes it possible (although not advised) to start posting heaps before
the connection has completed.