Add --verify option to spead2_send and spead2_recv
It puts random data (created from a linear congruential generator) into
the data stream, and verifies it on the receive side. This can help with
testing issues that are missed by unit tests e.g. corner cases that only
affect a small number of heaps, or that require the buffer to be
recycled before they show up.