Lock the correct lock when sending a heap.
It was locking the head mutex when it should have locked the tail mutex.
I discovered this when a race condition caused sending to stop, I'm
guessing because of a race condition on need_wakeup causing a wakeup to
get lost. But there would probably have been other entertaining race
conditions too.