Bring type stubs up to date
There were a whole lot of changes I'd missed updating for spead2 3.0.0.
The good news is that removing support for explicit loop arguments means
that the send stream constructors now have the same signatures between
sync and sync versions, which has allowed a lot of code duplication to
be eliminated.