Embed packet generator in each queue_item
This is probably a step backwards in the short term, but it paves the
way for having batches of heaps whose packets are interleaved.

There are some potential memory (and compute) savings in future around
having the max packet size be stored only once instead of in each packet
generator.