Move queue_item out of stream
This allows the forward references between stream and writer to be
inverted, so that writer is a complete type in send_stream.h. That is
going to become necessary to write async_send_heaps as a template
function.

To make it all work, the stream_config class (and dependencies) were
moved out to send_stream_config.h/cpp.