Fix mem_reader to stop gracefully
Previously it was directly calling `stop` on the stream. This is
equivalent to the user calling it, which is normally a hard stop e.g. if
it is a ring_stream, it will stop the ringbuffer without pushing any
incomplete heaps into it, or whatever flushing a subclass has
implemented in its `stop_received` overload.

Instead, use `add_packet_state.stop`, which is what's normally called
for a stop triggered by the stream itself (for example, a stop heap or
end-of-file in a pcap file), which indirectly calls `stop_received` on
the stream.