Redesign interaction of semaphores with the GIL
Previously it was a property of the semaphore type (a semaphore_gil
wrapper type would drop the GIL), but that doesn't allow for a semaphore
that's sometimes waited from C++ worker threads (no GIL drop) and
sometimes from Python (drop the GIL).

Instead, functions that need to block on a semaphore take a variadic
argument list, which is forwarded to the `semaphore_get` free function,
and py_common.h provides a version of `semaphore_get` that drops the
GIL.

In a few places sem.get() had to be replaced with semaphore_get(sem,
...) to get this feature into use. I believe these changes actually fix
bugs.