WIP on custom counters
Redefines stream_stats to hold a vector of counters and add
stream_stats_metadata to hold the names and types. The Python support is
still minimal (just backwards compatibility, no access to new features),
and there are a number of handy accessors that are missing, plus it
would help to have a better way to programmatically walk the counters.