Redesign deleter for memory_allocator::pointer
Previously it held a shared_ptr reference to the allocator, and called
back to the allocator to free the memory via a virtual function.
This had a few disadvantages:

- The allocator had to be kept alive (and referenced) even when its
  state wasn't needed for deletion (which was pretty much all the time).
- You had to have an allocator class, which made it lots of work to
  convert a regular unique_ptr to a memory_allocator::pointer.
- Any additional state had to be crammed into a void*, which had low
  safety (raw pointer AND untyped).

Replace it all by having the deleter type derive from std::function.
There is still backwards-compatible support for custom allocators that
override `free` and/or expect to be able to call `get_allocator` and
`get_user` on the deleter, but the allocators defined by spead2 do not
use it.