Add .values to Python StreamStats
Uses the new `make_value_iterator` in pybind11 2.8.0. I also had to
change the const_iterator to reference the value instead of copying it,
to work around a subtle typing bug in pybind11 (accessing the iterator
would return an rvalue reference to the value in the temporary pair,
rather than a copy).