Fix gdrapi_example.cu
It was updated but not tested (lack of hardware) when the memory
allocator API was changed. It broke because it tried to get the
gdrapi_memory_allocator::deleter from the pointer, but it was getting
the deleter for the memory pool.

Fixed by adding some API to memory_pool to allow the underlying deleter
to be retrieved.