Add multi-producer stop support for ringbuffers
Consumers can use now use `add_producer` and `remove_producer` to have
the ringbuffer stop only when the last producer is done.

This also changes `stop` to return a boolean indicating whether this is
the first time the ringbuffer was stopped, and fixes a minor issue where
repeated calls to `stop` would add more and more to the semaphores (only
really a problem for a pipe semaphore, which has a fairly bounded
capacity).