Proof of concept for memcpy_nontemporal_rw
This is a memcpy that uses non-temporal loads as well as non-temporal
stores. The goal is to see if this makes any performance improvement in
bandwidth-limited applications by reducing cache pollution.

It's definitely not ready to merge: it will probably fail to compile on
anything except x86 (and possibly only x86-64), and needs documentation
plus wiring up to command-line tools.