Fix bugs in reporting stats from spead2_recv
In the Python version it was printing some bogus fields because it was
using introspection on the object. Fix by making it using the newer
dictionary-like API.

In the C++ version, replace the hardcoded list of stats fields with the
map-like API. Also fix stats being printed in hex due to cout being left
in hex mode.