Improve accuracy of rate limiter
Previously it could be a bit off because wait times were rounded to
whole nanoseconds (or whatever the unit of high_resolution_clock is).
While the actual sleeps can only be this accurate, the residual errors
would accumulate and should eventually be corrected.

This is fixed by using a new precise_time class for internal time
calculations, that holds a time_point for the high resolution clock, and
a small correction term.