Make default constructor for precise_time well-defined
The default constructor for `duration` is defaulted, meaning it gives an
undefined value, while the default constructor for `time_point` is
defined to be the start of the epoch. The old code was relying on the
latter for `send_time` and `send_time_burst`, and the change to
`precise_time` was causing undefined behaviour. So the default is
changed to ensure the same default as for `time_point`.