Add support for pcap dumps in linux sll format
When packets are captured from the "any" device in tcpdump/wireshark,
the link type declared by the savefile is not Ethernet, but "Linux SLL",
a pseudo-protocol used by libpcap, also commonly referred to as "cooked
mode". This protocol uses its own header and thus needs to be parsed
differently.

This commit adds the required support to parse a UDP payload from a
Linux cooked-mode dump. When Ethernet frames are recorded in this
format, the Ethernet protocol is stored in the protocol field of the
sll frame, and the IP payload is stored directly as the payload of
the frame (i.e., the ethernet frame itself is not recorded).

Unit tests have been adjusted to test this new frame support.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>