Parse packets depending on the link type
The libpcap savefile states the device's data link type used to store
the packets. Depending on this we now parse packets as being Ethernet
frames or Linux cooked mode frames.

Note that the previous iteration of this code did not check that the
link type was Ethernet. A check has been added now to check that the
link type is one of the supported ones.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>