Define DLT_* constants ourselves
These constants were moved from pcap/bfp.h to pcap/dlt.h in 1.8.0, and
we build spead2 against different libpcap versions that cross that
boundary. Originally pcap/bfp.h included internally pcap/dlt.h, but that
might not always be the case in the future.

To avoid having to guess which file to include we define the constants
ourselves.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>