Store chunk_place_data inside the stream class
It's stored contiguously with the scratch space that it points at. This
reduces the need to allocate data on the fly, and should also minimise
the number of cache lines that need to be touched.

There is also partial support for the 'extra' field, but the data
written there is not actually copied.