WIP on shutdown rewrite
This is an attempt to simplify the reader shutdown path. Instead of
stream::stop blocking until the readers signal that their completion
handlers have all run, try to make it safe for the handlers to run after
the stream has shut down.

To do this, the queue_mutex is moved inside an object managed by
shared_ptr, so that the handlers can keep it alive even after the stream
is destroyed. This does end up requiring creation and destruction of a
shared_ptr for every handler invocation, which isn't ideal; perhaps
there is some way to transfer it down the handler chain (custom
executor?).

This also means that reader no longer need a specific `stop` member
function; they just stop things in their destructors, generally
automatically.