Start wiring up chunk_stream_group to Python
This already exposed a bunch of compilation issues that appear just
because templates are now being instantiated. chunk_ring_stream is
extended to take on the responsibility for making the chunk allocate and
ready functions, simplifying adjust_chunk_config and
adjust_group_config.

The Python wiring is incomplete, as it does not include
chunk_stream_group_config.