Make chunk_stream_group own the streams
Instead of the user creating streams which add themselves to the group
(and remove themselves on stop), have the group keep a fixed set of
streams.

This has the advantage that the user is not responsible for keeping
references to the streams to keep them alive / destroy them when
appropriate.