Fix a bug with unwanted sharing of stats between stream_config
Copying a stream_config would leave them pointing at the same
vector<stream_stat_config>, with correctness consequences if either copy
later adds more statistics. This happens as part of
chunk_stream::adjust_config.

This could also lead to race conditions if a config was modified after
constructing a stream with it, since it would affect and stream_stats
instances that it shared with.

The type has been changed to shared_ptr<const vector<...>>
throughout to prevent modifying shared data. This makes add_stat much
less efficient (it copies the vector every time) but that is not on the
critical path.