Remove proactive flushing in chunk_stream_group
I think it could potentially cause back-filled chunks (after a jump) to
be evicted before another stream had a chance to acquire them.

The proactive flushing was previously needed to ensure that if the
network stopped all the streams, the group's chunks were flushed before
the last call to data_ring->remove_producer so that the chunks could be
retrieved. This is now solved in stream_stop_received by detecting the
last stream stopping and doing this flush.