Reduce number of calls to async_flush_until
If multiple streams try to get the same chunk ID from a group at the
same time, it's possible that each of them would have issued an
async_flush_until to all the others, causing O(N^2) total calls. Improve
this by tracking what we flushed to previously.

Additionally, use this information to ensure that newly-added streams
also flush to this point. This fixes a potential deadlock if a new
stream is added once things are already running, and it sees a heap
that's inside the group's window but behind the flush_until point, and
then it never advances further.