Implement Sequence protocol for ChunkStreamRingGroup
It was partially implemented, but now it's fully implemented and
registered as a virtual subclass of collections.abc.Sequence.

As part of this, the C++ implementation of __iter__ was removed, in
favour of the mixin version from collections.abc. The latter is more
robust as it won't access undefined memory if the sequence is mutated
while iterating.