Fix error with very large chunk ID
A chunk ID close to 2^63 could lead to overflow bugs. Fix it by using
unsigned chunk IDs when dealing with head and tail chunk IDs. This does
require some more careful handling for determining whether a chunk ID is
behind the head, since it's a comparison of a signed and an unsigned
value.