Fix a shutdown hang
If the process exits without explicitly stopping anything, it's possible
deadlock if the threadpools used for the streams in a group were created
after the group. This happens because `atexit` stops the threadpools
before it stops the group.

Fix it by giving the exit_stopper treatment to
chunk_stream_group_member. Since the member is necessarily constructed
after the thread pool it uses, this ensures that the member (and hence
the whole group) is stopped before the thread pool.