Add detection of function multi-versioning in the compiler
This support is used so that the SSE2 version of memcpy_nontemporal is
selected only if the CPU supports SSE2 at runtime.

On x86-64 there is guaranteed to be runtime support, so this only
affects i386. Note that i386 is not currently tested. The real
motivation is that it lays the groundwork for adding more specialised
variants e.g. AVX, AVX-512, MOVDIR64B etc.