Use std::make_unique where appropriate
This simplifies the pattern of creating a unique_ptr holding newly
allocated memory. Since C++17 doesn't provide
std::make_unique_for_overwrite (important for performance when
allocating large arrays of raw storage space), add a home-brew
implementation in common_unique.h. This file isn't intended to be part
of the public API, so it is in the src directory rather than include
(and fortunately it happens not to be needed by any of the header files).