Speed up cibuildwheel builds
This is done by removing the need to compile or install anything from
boost. For quite a while, boost_system has been header-only (and only
indirectly used, via asio), so references to it are removed. That left
boost_program_options as the only library that spead2 depends on - and
only for command-line tools, not the wheel. So add
`--without-program-options` configure option to allow that dependency to
be skipped, and use it from the Python build.

That allowed the unpacked boost to be "installed" just with a symlink to
the include files rather than a proper installation.

Closes #145.