meson: use py instead of py.full_path()
It's not necessary to get the path to run the Python installation as an
executable.

Also use the detected Python to run get_version.py, so that it will work
even if there isn't a 'python3' in $PATH (particularly useful when
running pip install with a Python that isn't on $PATH).