meson: simplify cuda_dep
Using `dependency('cuda')` is only necessary if linking CUDA into a
project that isn't using nvcc, or if using extra libs like cublas.
Neither applies here, so I've replaced cuda_dep with an empty
dependency. Dependencies are still helpful because one can use a
disabler if CUDA is not found or is disabled by user options.