Simplify buffer_reader construction
It used to take a buffer_info by reference and move out of it, due to
limitations in the original implementation of emplace_reader (I think
involving pushing the work onto a worker thread). Perfect forwarding is
no longer an issue, and buffer_info is non-copyable so there is no need
to worry about an accidental copy.