Add AVX and AVX-512 non-temporal memcpy implementations
The implementation has been substantially changed, so that there is an
explicit ifunc resolver, instead of using GCC's function
multi-versioning. This allows AVX-512 to be more conditionally enabled,
and also makes it possible to test all the implementations in the unit
test.

The unrolling factor is now also easier to control, and is set to 16 for
the SSE2 implementation since there are signs that this can marginally
improve performance.