Make async_send_heap and async_send_heaps work with completion tokens
When using a completion token rather than a callback function, the
boolean return is no longer available. Unfortunately this requires
having two overloads.

Using this with boost::asio::use_future will greatly simplify the common
case of sending a heap then at a later point blocking until it is sent
or throwing if there was a problem.