Reorganise the tutorial filenames
There is now a global counter that matches section numbering in the
document, rather than numbering versions of the sender and receiver
separately (which made it hard to know which versions to use together).
The filenames also have a short description for what they're about.

Accidentally merged in a rename from chunk_size to heap_size.