Optimise memory use in for loop over stream
Heaps can use memory from a fixed-capacity memory pool, so we want to
restrict the number in flight at a time. When doing `h = stream.pop()`,
it's possible that the ringbuffer was previously full, and popping makes
space and unblocks a receive thread that then allocates a new heap. That
can happen before the memory in the old `h` is freed by the assignment
operator.

First clear the old heap before popping, to avoid this condition.