Make packet_generator::next_packet work in-place
Previously it returned a new vector, which was then assigned over a
previous vector. That means incrementally allocating memory for the new
vector, then freeing the original vector. It's more efficient to keep
reusing the same vector object, which quickly grows to the size it needs
to be, then doesn't require any further memory allocation.

This can improve throughput for 1472-byte packets by 20%.