CI: use curl for downloads instead of wget
The aarch64 manylinux builds have started failing with errors about not
trusting the certificate for www.tcpdump.org, and that doesn't seem to
be a problem for curl. I've also had issues with wget and certificates
in the past (https://github.com/pypa/manylinux/issues/1203), so I'm
ditching it and going back to curl. It might be flakier since curl
doesn't automatically retry if the connection drops partway through (it
has a --retry option but that only retries on HTTP status codes that
indicate a transient error).